requirements.txt
pandas
numpy
streamlit
stmol==0.0.9
py3Dmol
biotite
ipywidgets
ipython_genutils
biopython
streamlit_molstar
